Commit #5 : [version 1.2] CRUD `operation Delete`, `Optional`, `ResponseEntity` and others

+++ library1/src/main/java/io/company/library/LibraryRestController.java
@@ -1,9 +1,8 @@
 package io.company.library;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
 import java.util.Optional;
 
 @RestController
@@ -24,7 +23,7 @@ public class LibraryRestController {
     //CRUD: read, find one book by id
     @GetMapping("getBook")
     public Book findBookById(Long id){
-
+        //
         Optional<Book> bookFound = bookservice.findBookById(id);
         if (bookFound.isPresent()) return  bookFound.get();
 
@@ -34,24 +33,35 @@ public class LibraryRestController {
     //CRUD: create
     @PostMapping(path="addBook", consumes = "application/JSON")
     public Book addBook(@RequestBody Book book){
-
-
+        //
         Book bookCreated = bookservice.createBook(book);
-
         return bookCreated ;
     }
 
     //CRUD: delete
     @DeleteMapping("deleteBook")
-    public Book deleteBook (Long id){
+    public ResponseEntity<Book> deleteBook (@RequestParam Long id) {
+        //
+        Optional<Book> bookFound  = bookservice.findBookById(id);
+        boolean isBook = bookFound.isPresent();
+        if(isBook) {
+            bookservice.deleteBookById(id);
+            return  ResponseEntity.accepted().body(bookFound.get());
+        } else return ResponseEntity.accepted().body(null);
 
-        bookservice.deleteBookById(id);
+
+    }
+
+    //CRUD: update
+    public ResponseEntity<Book> updateBook (@RequestBody Book bookToUpdate){
+        //find book by id >> bookFound
+        // if exists .... let s compare
+        // field by field
+        // bookToUpdate to bookFound
+        //bookservice.updateBook():
 
         return null;
     }
-
-
-    //CRUD: update
 
 
 }
